BAARC-86: Move logic for getting settings to helper functions and cmd package
