+++ config.yaml
@@ -80,7 +80,6 @@ blocktx:
       recordRetentionDays: 14 # number of days for which data is kept in the storage before it can be deleted
       executionIntervalHours: 24
   profilerAddr: localhost:9993 # address to start profiler server on
-  startingBlockHeight: 100 # starting block height for blocktx to start from. blocktx will not request blocks lower than this height
 
 broadcaster:
   apiURL: http://localhost:9090 # api url for broadcaster to connect to
